fix: null Inn employment error in deal db

+++ application/src/test/java/com/application/ApplicationTests.java
@@ -9,6 +9,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
@@ -20,7 +21,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @AutoConfigureMockMvc
 class ApplicationTests {
 	private final LoanApplicationRequestDTO validRequest = LoanApplicationRequestDTO.builder()
-			.amount(1000000.0)
+			.amount(BigDecimal.valueOf(1000000.0))
 			.term(24)
 			.firstName("Ivan")
 			.lastName("Ivanov")
@@ -32,7 +33,7 @@ class ApplicationTests {
 			.build();
 
 	private final LoanApplicationRequestDTO invalidRequest = LoanApplicationRequestDTO.builder()
-			.amount(-200.0)
+			.amount(BigDecimal.valueOf(-200.0))
 			.term(24)
 			.firstName("Ivan")
 			.lastName("Ivanov")
